+++ app/lib/models/items/video.dart
@@ -42,7 +42,8 @@ class VideoPartItem extends PartItem {
       "url": url,
       "source": EnumToString.convertToString(source),
       "name": name,
-      "description": description
+      "description": description,
+      "points": points
     };
   }